+++ src/layouts/Stub.vue
@@ -1,0 +1,13 @@
+<template>
+  <div class="stub">
+    This is stub layout
+    <slot />
+  </div>
+</template>
+
+<script lang="ts">
+import { Component, Vue } from 'vue-property-decorator';
+
+@Component
+export default class StubLayout extends Vue {}
+</script>
