+++ src/views/TechnicalWork.vue
@@ -1,0 +1,12 @@
+<template>
+  <div>Technical Work in Progress</div>
+</template>
+
+<script lang="ts">
+import { Component, Mixins } from 'vue-property-decorator';
+
+import TranslationMixin from '@/components/mixins/TranslationMixin';
+
+@Component
+export default class TechnicalWork extends Mixins(TranslationMixin) {}
+</script>
